pyproject.toml
[tool.poetry]
name = "mmc" #"Multi-Modal Comparators"
version = "0.1.0"
description = "Unified API to facilitate usage of pre-trained \"perceptor\" models, a la CLIP"
authors = ["David Marx <david.marx84@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
clip = {git = "https://github.com/openai/CLIP", branch = "main"}
loguru = "^0.6.0"
Pillow = "^7.1.2" #"^9.1.0"
#torch = "^1.11.0"
#torchvision = "^0.12.0"
#torchaudio = "^0.11.0"
kornia = "^0.6.4"
open-clip-torch = "^1.0.1"
declip = {git = "https://github.com/pytti-tools/DeCLIP", branch = "installable"}
kelip = {git = "https://github.com/navervision/KELIP.git", branch = "master"}
sentence-transformers = "^2.2.0"
napm="^0.2.0"
timm="^0.5.4"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.1"
poethepoet = "^0.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# https://github.com/nat-n/poethepoet
[tool.poe.tasks]
napm_installs = { "script" = "mmc.napm_installs:all" }
